backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sse-starlette==2.1.0
google-generativeai==0.8.0
httpx==0.27.0
python-dotenv==1.0.1
groq==0.12.0
pdfplumber==0.11.4
python-multipart==0.0.9
